Field isolation into distinct modules.
