+++ marrow/mongo/core/field/md.py
@@ -6,7 +6,7 @@ from misaka import Markdown, HtmlRenderer  # , SmartyPants
 from misaka import HTML_ESCAPE, HTML_HARD_WRAP
 from misaka import EXT_FENCED_CODE, EXT_NO_INTRA_EMPHASIS, EXT_AUTOLINK, EXT_SPACE_HEADERS, EXT_STRIKETHROUGH, EXT_SUPERSCRIPT
 
-from .base import String
+from .string import String
 from ....schema.compat import unicode, py3
